pyproject.toml
[tool.poetry]
name = "geotortoise"
version = "0.1.1"
description = "Tortoise-orm with geospatial operations "
authors = ["PhotoPills Team <it@photopills.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tortoise-orm = "^0.17.2"
Shapely = "^1.7.1"
asyncpg = "^0.22.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
asynctest = "^0.13.0"
black = "^21.4b0"
flake8 = "^3.9.1"
pre-commit = "^2.12.1"
isort = "^5.8.0"
aerich = "^0.5.3"
pytest-aiohttp = "^0.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
